+++ DHARMa/R/tests.R
@@ -86,7 +86,7 @@ testBivariateUniformity <- function(simulationOutput, alternative = c("two.sided
 #'
 #' The p-values of the intercept and splines are combined into a total p-value via Benjamini & Hochberg adjustment to control the FDR.
 #'
-#' Predictor needs to be a variable in your environment (e.g. predictor = data$predictor). For more details and a more flexible syntax for predictors, see the help of the argument `form` in [plotResiduals].
+#' Predictor can be a formula (e.g. predictor = ~predictor), in which case NAs are handled automatically (recommended). Predictor can also be a variable in your environment (e.g. predictor = data$predictor), but then you need to remove rows that were excluded by the model due to NAs by hand. For more details and a more flexible syntax for predictors, see the help of the argument `form` in [plotResiduals].
 #'
 #' When plotting (plot = TRUE), the shaded gray areas indicate 95% confidence intervals of the quantile estimates (1.96 * standard error).
 #'
@@ -104,6 +104,8 @@ testQuantiles <- function(simulationOutput, predictor = NULL, rank = TRUE,
 
     simulationOutput = ensureDHARMa(simulationOutput, convert = T)
     res = simulationOutput$scaledResiduals
+
+    if(inherits(predictor, "formula")) predictor = getFormulaPredictors(simulationOutput, predictor)[[1]]
     pred = ensurePredictor(simulationOutput, predictor)
     if (rank == TRUE) pred = rankTransform(pred)
 
@@ -336,7 +338,7 @@ testOutliers <- function(simulationOutput, alternative = c("two.sided", "greater
 #' This function tests if there are problems in a res ~ group structure. It performs two tests: test for within-group uniformity, and test for between-group homogeneity of variances.
 #'
 #' @param simulationOutput an object of class DHARMa, either created via [simulateResiduals] for supported models or by [createDHARMa] for simulations created outside DHARMa, or a supported model. Providing a supported model directly is discouraged, because simulation settings cannot be changed in this case.
-#' @param catPred a categorical predictor with the same dimensions as the residuals in simulationOutput.
+#' @param catPred a categorical predictor with the same dimensions as the residuals in simulationOutput. If specified as formula, e.g. catPred = ~group, NAs are handled automatically (recommended).
 #' @param quantiles whether to draw the quantile lines.
 #' @param plot if TRUE, the function will create an additional plot.
 #' @param ... additional arguments to boxplot.
@@ -357,6 +359,12 @@ testCategorical <- function(simulationOutput, catPred,
 
   a = list(...)
   a$xlab = paste(checkDots("xlab", deparse(substitute(catPred)), ...), "(catPred)")
+
+  if(!inherits(catPred, "formula")) {
+    catPred = catPred # old default
+  } else {
+    catPred = getFormulaPredictors(simulationOutput, formula = catPred)[[1]] # formula syntax
+  }
 
   catPred = as.factor(catPred)
   out = list()
@@ -625,7 +633,7 @@ testGeneric <- function(simulationOutput, summary, alternative = c("two.sided", 
 #' This function performs a standard test for temporal autocorrelation on the simulated residuals.
 #'
 #' @param simulationOutput an object of class DHARMa, either created via [simulateResiduals] for supported models or by [createDHARMa] for simulations created outside DHARMa, or a supported model. Providing a supported model directly is discouraged, because simulation settings cannot be changed in this case.
-#' @param time the time, in the same order as the data points.
+#' @param time time variable in the same order as the datapoints and with the same length as the residuals in simulationOutput. If specified as formula, e.g. time = ~time, NAs are handled automatically (recommended).
 #' @param alternative a character string specifying whether the test should test if observations are "greater", "less" or "two.sided" compared to the simulated null hypothesis.
 #' @param plot whether to plot output.
 #' @details The function performs a Durbin-Watson test on the uniformly scaled residuals, and plots the residuals against time. The DB test was originally designed for normal residuals. In simulations, I didn't see a problem with this setting though. The alternative is to transform the uniform residuals to normal residuals and perform the DB test on those.
@@ -654,9 +662,6 @@ testTemporalAutocorrelation <- function(simulationOutput, time, alternative = c(
 
   simulationOutput = ensureDHARMa(simulationOutput, convert = T)
 
-  # actually not sure if this is neccessary for dwtest, but seems better to aggregate
-  if(any(duplicated(time))) stop("testing for temporal autocorrelation requires unique time values - if you have several observations per time value, either use the recalculateResiduals function to aggregate residuals per time step, or extract the residuals from the fitted object, and plot / test each of them independently for temporally repeated subgroups (typical choices would be location / subject etc.). Note that the latter must be done by hand, outside testTemporalAutocorrelation.")
-
   alternative <- match.arg(alternative)
 
   if(is.null(time)){
@@ -664,8 +669,16 @@ testTemporalAutocorrelation <- function(simulationOutput, time, alternative = c(
     message("DHARMa::testTemporalAutocorrelation - no time argument provided, using random times for each data point")
   }
 
-  # To avoid Issue #190
-  if (length(time) != length(residuals(simulationOutput))) stop("Dimensions of time don't match the dimension of the residuals")
+  if(!inherits(time, "formula")){
+    # old default
+    # To avoid Issue #190
+    if (length(time) != length(residuals(simulationOutput))) stop("Dimensions of time don't match the dimension of the residuals. Remove rows with NAs or use formula syntax for group to handle NAs automatically.")
+  } else {
+    time = getFormulaPredictors(simulationOutput, formula = time)[[1]] # formula syntax
+  }
+
+  # actually not sure if this is neccessary for dwtest, but seems better to aggregate
+  if(any(duplicated(time))) stop("testing for temporal autocorrelation requires unique time values - if you have several observations per time value, either use the recalculateResiduals function to aggregate residuals per time step, or extract the residuals from the fitted object, and plot / test each of them independently for temporally repeated subgroups (typical choices would be location / subject etc.). Note that the latter must be done by hand, outside testTemporalAutocorrelation.")
 
   out = lmtest::dwtest(simulationOutput$scaledResiduals ~ 1, order.by = time, alternative = alternative)
 
@@ -735,16 +748,33 @@ testSpatialAutocorrelation <- function(simulationOutput, x = NULL, y  = NULL, di
 
   # Assertions
 
+  if( (!is.null(x) | !is.null(y)) & !is.null(distMat) ){
+    if(isTRUE(plot)) {message("Both coordinates and distMat provided, calculations will be done based on the distance matrix, coordinates will only be used for plotting.")}
+
+    else { # set x,y to NULL so they don't cause errors if wrong dimensions
+      x = NULL
+      y = NULL
+      message("Both coordinates and distMat provided, calculations will be done based on the distance matrix, coordinates will be ignored.")
+      }
+  }
+
+  if( (is.null(x) | is.null(y)) & is.null(distMat) ) stop("You need to provide either x,y coordinates or a distMatrix.")
+
+  # check if x,y are formulas
+  if(!inherits(x, "formula") | !inherits(y, "formula")) {
+    # old default
+    # To avoid Issue #190
+    if (is.null(distMat) & !is.null(x) & length(x) != length(residuals(simulationOutput)) | is.null(distMat) & !is.null(y) & length(y) != length(residuals(simulationOutput))) stop("Dimensions of x / y coordinates don't match the dimension of the residuals. Remove rows with NAs or specify x,y as formula to handle NAs automatically.")
+
+    if(inherits(x, "formula") != inherits(y, "formula")) stop("Please specify x and y the same way, either both as formula (e.g.x = ~x) or both as vector (e.g. x = data$x).")
+  } else {
+    # formula syntax
+    x = getFormulaPredictors(simulationOutput, formula = x)[[1]]
+    y = getFormulaPredictors(simulationOutput, formula = y)[[1]]
+  }
+
+
   if(any(duplicated(cbind(x,y)))) stop("Testing for spatial autocorrelation requires unique x,y values - if you have several observations per location, either use the recalculateResiduals function to aggregate residuals per location, or extract the residuals from the fitted object, and plot / test each of them independently for spatially repeated subgroups (a typical scenario would be repeated spatial observation, in which case one could plot / test each time step separately for temporal autocorrelation). Note that the latter must be done by hand, outside testSpatialAutocorrelation.")
-
-  if( (!is.null(x) | !is.null(y)) & !is.null(distMat) ) message("Both coordinates and distMat provided, calculations will be done based on the distance matrix, coordinates will only be used for plotting.")
-
-  if( (is.null(x) | is.null(y)) & is.null(distMat) ) stop("You need to provide either x,y, coordinates or a distMatrix.")
-
-  if(is.null(distMat) & (length(x) != length(residuals(simulationOutput)) | length(y) != length(residuals(simulationOutput))))
-
-  # To avoid Issue #190
-  if (!is.null(x) & length(x) != length(residuals(simulationOutput)) | !is.null(y) & length(y) != length(residuals(simulationOutput))) stop("Dimensions of x / y coordinates don't match the dimension of the residuals.")
 
   # if not provided, create distance matrix based on x and y
   if(is.null(distMat)) distMat <- as.matrix(dist(cbind(x, y)))