+++ DHARMa/inst/examples/wrappersHelp.R
@@ -17,3 +17,6 @@ x = getSimulations(fittedModel, nsim = 2, type = "refit")
 getRefit(fittedModel, x[[1]])
 
 getRefit(fittedModel, getObservedResponse(fittedModel))
+
+# get data frame that was used to fit the model
+getData(fittedModel)